+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Fixed
+
+- `Interval::H8` now formats as `8h` instead of `1m`
+- re-sync any 8h candles written before this fix; they were stored under the `1m` interval key
+
 ## [0.2.0] - 2026-03-21
 
 ### Changed